Pom changes (#53)
* Added a new POM file that encapsulates dependencies on all five modules

* Made com.microsoft.azure:azure-client-authentication scope to test

* Added some dependency management features

* Modified dependencyManagement

* Cleaned up webkey

* Cleaned up cryptography

* Cleaned up core

* Added in used but undeclared dependencies

* Removed dependency on commons-io, which appears unneeded

* Cleaned up unused dependencies in extensions and fixed minor typographical problems

* Added undeclared, but used, dependencies to extensions

* Added some more  package versions to the POMP

* Extracted shared versions to parent pom

* Added explicit dependencies to extensions and fixed some minor errors

* Fixed some indentation differences

* Fixed IT bug:

* Updated all versions to 1.1.1

* fixed dependency version error

* Updated integration-test path
