+++ azure-keyvault-complete/pom.xml
@@ -1,0 +1,72 @@
+<!-- Copyright (c) Microsoft Corporation. All rights reserved. Licensed under 
+the MIT License. See License.txt in the project root for license information. -->
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+  
+  <parent>
+		<groupId>com.microsoft.azure</groupId>
+		<artifactId>azure-keyvault-parent</artifactId>
+		<version>1.1.1</version>
+		<relativePath>../pom.xml</relativePath>
+	</parent>
+  
+  <groupId>com.microsoft.azure</groupId>
+  <artifactId>azure-keyvault-complete</artifactId>
+  <version>1.1.1</version>
+  <packaging>pom</packaging>
+  
+  <licenses>
+    <license>
+      <name>The MIT License (MIT)</name>
+      <url>http://opensource.org/licenses/MIT</url>
+      <distribution>repo</distribution>
+    </license>
+  </licenses>
+  
+  <name>Microsoft Azure Key Vault SDK Complete</name>
+  <description>This encapsulate all modules of the Microsoft Azure Key Vault SDK.</description>
+  <url>https://github.com/Azure/azure-keyvault-java</url>
+  
+  <properties>
+    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+    <legal><![CDATA[[INFO] Any downloads listed may be third party software.  Microsoft grants you no rights for third party software.]]></legal>
+  </properties>
+  
+  <scm>
+    <url>scm:git:https://github.com/Azure/azure-keyvault-java</url>
+    <connection>scm:git:git@github.com:Azure/azure-keyvault-java.git</connection>
+    <tag>HEAD</tag>
+  </scm>
+  
+  <developers>
+    <developer>
+      <id>microsoft</id>
+      <name>Microsoft</name>
+    </developer>
+  </developers>
+  
+  <dependencies>
+    <dependency>
+      <groupId>com.microsoft.azure</groupId>
+      <artifactId>azure-keyvault</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>com.microsoft.azure</groupId>
+      <artifactId>azure-keyvault-core</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>com.microsoft.azure</groupId>
+      <artifactId>azure-keyvault-webkey</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>com.microsoft.azure</groupId>
+      <artifactId>azure-keyvault-extensions</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>com.microsoft.azure</groupId>
+      <artifactId>azure-keyvault-cryptography</artifactId>
+    </dependency>
+  </dependencies>
+  
+  </project>
