+++ azure-keyvault-extensions/pom.xml
@@ -8,7 +8,7 @@
   <parent>
     <groupId>com.microsoft.azure</groupId>
     <artifactId>azure-keyvault-parent</artifactId>
-    <version>1.1.0</version>
+    <version>1.1.1</version>
     <relativePath>../pom.xml</relativePath>
   </parent>
 
@@ -47,14 +47,78 @@
   </developers>
 
   <dependencies>
+    <!-- COMPILE DEPENDENCIES -->
+    <!-- Microsoft Azure KeyVault Dependencies -->
+    <dependency>
+    	<groupId>com.microsoft.azure</groupId>
+    	<artifactId>azure-keyvault-webkey</artifactId>
+    </dependency>
+    <dependency>
+    	<groupId>com.microsoft.azure</groupId>
+    	<artifactId>azure-keyvault-core</artifactId>
+    </dependency>
+    <dependency>
+    	<groupId>com.microsoft.azure</groupId>
+    	<artifactId>azure-keyvault-cryptography</artifactId>
+    </dependency>
+    <dependency>
+    	<groupId>com.microsoft.azure</groupId>
+    	<artifactId>azure-keyvault</artifactId>
+      <version>1.1.1</version>
+    </dependency>
+    
+    <!-- Other Microsoft Azure Dependencies -->
     <dependency>
       <groupId>com.microsoft.azure</groupId>
       <artifactId>azure-client-runtime</artifactId>
     </dependency>
     <dependency>
+      <groupId>com.microsoft.azure</groupId>
+      <artifactId>azure-mgmt-resources</artifactId>
+    </dependency>
+    
+    <!-- Other Microsoft Dependencies -->
+    <dependency>
+      <groupId>com.microsoft.rest</groupId>
+      <artifactId>client-runtime</artifactId>
+    </dependency>
+ 
+    <!-- Other Dependencies -->
+    <dependency>
+    	<groupId>com.google.guava</groupId>
+    	<artifactId>guava</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.apache.commons</groupId>
+      <artifactId>commons-lang3</artifactId>
+    </dependency>
+    <dependency>
       <groupId>commons-codec</groupId>
       <artifactId>commons-codec</artifactId>
     </dependency>
+    
+    
+    <!-- TEST DEPENDENCIES -->
+    <!-- Other Microsoft Azure Dependencies -->
+    <dependency>
+      <groupId>com.microsoft.azure</groupId>
+      <artifactId>azure-mgmt-resources</artifactId>
+      <version>1.3.1-SNAPSHOT</version>
+      <type>test-jar</type>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+			<groupId>com.microsoft.azure</groupId>
+			<artifactId>adal4j</artifactId>
+			<scope>test</scope>
+		</dependency>
+    <dependency>
+			<groupId>com.microsoft.azure</groupId>
+			<artifactId>azure</artifactId>
+			<scope>test</scope>
+		</dependency>
+    
+    <!-- Other Dependencies -->
     <dependency>
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
@@ -64,40 +128,6 @@
       <groupId>org.bouncycastle</groupId>
       <artifactId>bcprov-jdk15on</artifactId>
       <scope>test</scope>
-      <version>1.54</version>
-    </dependency>
-    <dependency>
-    	<groupId>com.microsoft.azure</groupId>
-    	<artifactId>azure-keyvault-core</artifactId>
-    	<version>1.1.0</version>
-    </dependency>
-    <dependency>
-    	<groupId>com.microsoft.azure</groupId>
-    	<artifactId>azure-keyvault-cryptography</artifactId>
-    	<version>1.1.1</version>
-    </dependency>
-    <dependency>
-    	<groupId>com.microsoft.azure</groupId>
-    	<artifactId>azure-keyvault</artifactId>
-    	<version>1.1.1</version>
-    </dependency>
-    <dependency>
-    	<groupId>com.microsoft.azure</groupId>
-    	<artifactId>azure-keyvault-webkey</artifactId>
-    	<version>1.1.0</version>
-    </dependency>
-      		<dependency>
-			<groupId>com.microsoft.azure</groupId>
-			<artifactId>azure</artifactId>
-			<version>1.3.0</version>
-			<scope>test</scope>
-		</dependency>
-      <dependency>
-      <groupId>com.microsoft.azure</groupId>
-      <artifactId>azure-mgmt-resources</artifactId>
-      <version>1.3.1-SNAPSHOT</version>
-      <type>test-jar</type>
-      <scope>test</scope>
     </dependency>
     <dependency>
     	<groupId>org.mockito</groupId>
@@ -105,17 +135,8 @@
     	<version>1.10.19</version>
     	<scope>test</scope>
     </dependency>
-    <dependency>
-    	<groupId>com.microsoft.azure</groupId>
-    	<artifactId>adal4j</artifactId>
-    	<scope>test</scope>
-    </dependency>
-    <dependency>
-    	<groupId>com.google.guava</groupId>
-    	<artifactId>guava</artifactId>
-    	<version>20.0</version>
-    </dependency>
   </dependencies>
+  
   <build>
     <plugins>
       <plugin>
