+++ azure-keyvault-webkey/pom.xml
@@ -6,12 +6,12 @@
 	<parent>
 		<groupId>com.microsoft.azure</groupId>
 		<artifactId>azure-keyvault-parent</artifactId>
-		<version>1.1.0</version>
+		<version>1.1.1</version>
 		<relativePath>../pom.xml</relativePath>
 	</parent>
 
 	<artifactId>azure-keyvault-webkey</artifactId>
-  <version>1.1.0</version>
+  <version>1.1.1</version>
 	<packaging>jar</packaging>
 
 	<name>Microsoft Azure SDK for Key Vault WebKey</name>
@@ -45,10 +45,22 @@
 	</developers>
 
 	<dependencies>
+    <!-- COMPILE DEPENDENCIES -->
+    <!-- Jackson Dependencies -->
 		<dependency>
 			<groupId>com.fasterxml.jackson.core</groupId>
 			<artifactId>jackson-databind</artifactId>
 		</dependency>
+    <dependency>
+			<groupId>com.fasterxml.jackson.core</groupId>
+			<artifactId>jackson-core</artifactId>
+		</dependency>
+    <dependency>
+			<groupId>com.fasterxml.jackson.core</groupId>
+			<artifactId>jackson-annotations</artifactId>
+		</dependency>
+    
+    <!-- Other Dependencies -->
 		<dependency>
 			<groupId>commons-codec</groupId>
 			<artifactId>commons-codec</artifactId>
@@ -56,24 +68,17 @@
 		<dependency>
 			<groupId>com.google.guava</groupId>
 			<artifactId>guava</artifactId>
-			</dependency>
-		<dependency>
-			<groupId>com.microsoft.rest</groupId>
-			<artifactId>client-runtime</artifactId>
-			<version>1.3.0</version>
-		</dependency>
-		<dependency>
-			<groupId>com.microsoft.azure</groupId>
-			<artifactId>azure-client-authentication</artifactId>
-			<version>1.3.0</version>
-		</dependency>
-		<!-- Test dependencies -->
+    </dependency>
+    
+		<!-- TEST DEPENDENCIES -->
 		<dependency>
 			<groupId>junit</groupId>
 			<artifactId>junit</artifactId>
 			<scope>test</scope>
 		</dependency>
 	</dependencies>
+  
+
 	<build>
 		<plugins>
 			<plugin>
