+++ azure-keyvault/pom.xml
@@ -6,12 +6,12 @@ the MIT License. See License.txt in the project root for license information. --
 	<parent>
 		<groupId>com.microsoft.azure</groupId>
 		<artifactId>azure-keyvault-parent</artifactId>
-		<version>1.1.0</version>
+		<version>1.1.1</version>
 		<relativePath>../pom.xml</relativePath>
 	</parent>
 
 	<artifactId>azure-keyvault</artifactId>
-  <version>1.1.1</version>
+	<version>1.1.1</version>
 	<packaging>jar</packaging>
 
 	<name>Microsoft Azure SDK for Key Vault</name>
@@ -59,80 +59,151 @@ the MIT License. See License.txt in the project root for license information. --
 	</repositories>
 
 	<dependencies>
-		<dependency>
-			<groupId>com.microsoft.azure</groupId>
-			<artifactId>azure</artifactId>
-			<version>1.3.0</version>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>com.microsoft.azure</groupId>
-			<artifactId>azure-client-runtime</artifactId>
-			<version>[1.1.0,2.0.0)</version>
-		</dependency>
-		<dependency>
-			<groupId>com.microsoft.azure</groupId>
-			<artifactId>azure-client-authentication</artifactId>
-			<version>[1.1.0,2.0.0)</version>
-		</dependency>
-		<dependency>
+    <!-- COMPILE DEPENDENCIES -->
+    <!-- Microsoft Azure KeyVault SDK Dependencies -->
+    <dependency>
 			<groupId>com.microsoft.azure</groupId>
 			<artifactId>azure-keyvault-cryptography</artifactId>
-			<version>1.1.1</version>
-			<scope>system</scope>
-			<systemPath>${basedir}/../azure-keyvault-cryptography/target/azure-keyvault-cryptography-1.1.1.jar</systemPath>
-		</dependency>
-		<dependency>
-			<groupId>com.microsoft.azure</groupId>
-			<artifactId>azure-keyvault-core</artifactId>
-			<version>1.1.0</version>
-			<scope>system</scope>
-			<systemPath>${basedir}/../azure-keyvault-core/target/azure-keyvault-core-1.1.0.jar</systemPath>
 		</dependency>
 		<dependency>
 			<groupId>com.microsoft.azure</groupId>
 			<artifactId>azure-keyvault-webkey</artifactId>
-			<version>1.1.0</version>
 		</dependency>
-		<!-- Test dependencies -->
-		<dependency>
-			<groupId>junit</groupId>
-			<artifactId>junit</artifactId>
+    
+    <!-- Other Microsoft Azure Dependencies -->
+    <dependency>
+			<groupId>com.microsoft.azure</groupId>
+			<artifactId>azure-client-runtime</artifactId>
+		</dependency>
+    
+    <!-- Other Microsoft Dependencies -->
+    <dependency>
+      <groupId>com.microsoft.rest</groupId>
+      <artifactId>client-runtime</artifactId>
+    </dependency>
+    
+    <!-- Jackson Dependencies -->
+    <dependency>
+			<groupId>com.fasterxml.jackson.core</groupId>
+			<artifactId>jackson-databind</artifactId>
+		</dependency>
+    <dependency>
+			<groupId>com.fasterxml.jackson.core</groupId>
+			<artifactId>jackson-annotations</artifactId>
+		</dependency>
+    <dependency>
+			<groupId>com.fasterxml.jackson.core</groupId>
+			<artifactId>jackson-core</artifactId>
+		</dependency>
+    
+    <!-- Other Dependencies -->
+    <dependency>
+      <groupId>commons-codec</groupId>
+      <artifactId>commons-codec</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>com.google.guava</groupId>
+      <artifactId>guava</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.apache.commons</groupId>
+      <artifactId>commons-lang3</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>joda-time</groupId>
+      <artifactId>joda-time</artifactId>
+      <version>2.1</version>
+    </dependency>
+    <dependency>
+      <groupId>com.squareup.retrofit2</groupId>
+      <artifactId>retrofit</artifactId>
+      <version>2.4.0</version>
+    </dependency>
+    <dependency>
+      <groupId>com.squareup.okhttp3</groupId>
+      <artifactId>okhttp</artifactId>
+      <version>3.10.0</version>
+    </dependency>
+    <dependency>
+      <groupId>com.squareup.okio</groupId>
+      <artifactId>okio</artifactId>
+      <version>1.14.0</version>
+    </dependency>
+    <dependency>
+      <groupId>io.reactivex</groupId>
+      <artifactId>rxjava</artifactId>
+    </dependency>
+    
+    <!-- TEST DEPENDENCIES -->
+    <!-- Other Microsoft Azure Dependencies -->
+    <dependency>
+      <groupId>com.microsoft.azure</groupId>
+      <artifactId>azure-mgmt-storage</artifactId>
+      <version>1.3.0</version>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+			<groupId>com.microsoft.azure</groupId>
+			<artifactId>azure-mgmt-graph-rbac</artifactId>
+			<version>1.3.0</version>
 			<scope>test</scope>
 		</dependency>
-		<dependency>
+    <dependency>
+      <groupId>com.microsoft.azure</groupId>
+       <artifactId>azure-mgmt-resources</artifactId>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
 			<groupId>com.microsoft.azure</groupId>
 			<artifactId>azure-mgmt-keyvault</artifactId>
-			<version>1.11.1</version>
 			<scope>test</scope>
+			<exclusions>
+				<exclusion>
+					<groupId>com.microsoft.azure</groupId>
+					<artifactId>azure-keyvault</artifactId>
+				</exclusion>
+			</exclusions>
 		</dependency>
-		<dependency>
+    <dependency>
 			<groupId>com.microsoft.azure</groupId>
 			<artifactId>azure-mgmt-resources</artifactId>
 			<version>1.3.1-SNAPSHOT</version>
 			<type>test-jar</type>
 			<scope>test</scope>
 		</dependency>
-		<dependency>
+    <dependency>
 			<groupId>com.microsoft.azure</groupId>
 			<artifactId>adal4j</artifactId>
 			<scope>test</scope>
 		</dependency>
-		<dependency>
-			<groupId>org.bouncycastle</groupId>
-			<artifactId>bcprov-jdk15on</artifactId>
-			<version>1.54</version>
+    <dependency>
+      <groupId>com.microsoft.azure</groupId>
+      <artifactId>azure-storage</artifactId>
+      <version>4.4.0</version>
+      <scope>test</scope>
+    </dependency>
+    
+    <!-- Other Dependencies -->
+    <dependency>
+      <groupId>com.google.code.gson</groupId>
+      <artifactId>gson</artifactId>
+      <version>2.2.4</version>
+      <scope>test</scope>
+    </dependency>
+    		<dependency>
+			<groupId>com.microsoft.azure</groupId>
+			<artifactId>azure-client-authentication</artifactId>
+			<version>[1.1.0,2.0.0)</version>
+      <scope>test</scope>
+		</dependency>
+    <dependency>
+			<groupId>junit</groupId>
+			<artifactId>junit</artifactId>
 			<scope>test</scope>
 		</dependency>
-		<!-- https://mvnrepository.com/artifact/commons-io/commons-io -->
-		<dependency>
-			<groupId>commons-io</groupId>
-			<artifactId>commons-io</artifactId>
-			<version>2.6</version>
-			<scope>test</scope>
-		</dependency>
-
-	</dependencies>
+	</dependencies>	
+		
+		
 
 	<build>
 		<plugins>
@@ -170,6 +241,16 @@ the MIT License. See License.txt in the project root for license information. --
 					<source>1.7</source>
 					<target>1.7</target>
 				</configuration>
+			</plugin>
+			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-failsafe-plugin</artifactId>
+				<version>2.22.0</version>
+          			<configuration>
+			            <additionalClasspathElements>
+			              <additionalClasspathElement>${basedir}/azure-keyvault/target/azure-keyvault-${project.version}.jar</additionalClasspathElement>
+			            </additionalClasspathElements>
+			          </configuration>
 			</plugin>
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
