+++ pom.xml
@@ -6,7 +6,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <groupId>com.microsoft.azure</groupId>
-  <version>1.1.0</version>
+  <version>1.1.1</version>
   <artifactId>azure-keyvault-parent</artifactId>
   <packaging>pom</packaging>
 
@@ -31,6 +31,11 @@
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <legal><![CDATA[[INFO] Any downloads listed may be third party software.  Microsoft grants you no rights for third party software.]]></legal>
+    
+    <!-- Version properties -->
+    <jackson.version>2.9.6</jackson.version>
+    
+    <!-- Testing properties -->
     <testMode>playback</testMode>
     <playbackServerPort>11080</playbackServerPort>
     <alternativePlaybackServerPort>11081</alternativePlaybackServerPort>
@@ -81,28 +86,87 @@
 
   <dependencyManagement>
     <dependencies>
-	<!-- https://mvnrepository.com/artifact/com.microsoft.azure/azure-mgmt-keyvault -->
-	<dependency>
-    		<groupId>com.microsoft.azure</groupId>
-    		<artifactId>azure-mgmt-keyvault</artifactId>
-    		<version>1.11.1</version>
-	</dependency>
+      <!-- Azure KeyVault SDK Dependencies -->
+      <dependency>
+        <groupId>com.microsoft.azure</groupId>
+        <artifactId>azure-keyvault-webkey</artifactId>
+        <version>1.1.1</version>
+      </dependency>
+      <dependency>
+        <groupId>com.microsoft.azure</groupId>
+        <artifactId>azure-keyvault-cryptography</artifactId>
+        <version>1.1.1</version>
+      </dependency>
+      <dependency>
+        <groupId>com.microsoft.azure</groupId>
+        <artifactId>azure-keyvault-core</artifactId>
+        <version>1.1.1</version>
+      </dependency>
+      <dependency>
+      <groupId>com.microsoft.azure</groupId>
+      <artifactId>azure-keyvault</artifactId>
+      <version>1.1.1</version>
+    </dependency>
+      <dependency>
+        <groupId>com.microsoft.azure</groupId>
+        <artifactId>azure-keyvault-extensions</artifactId>
+        <version>1.1.1</version>
+      </dependency>
+      
+      <!-- Other Microsoft Azure Dependencies -->
+      <!-- https://mvnrepository.com/artifact/com.microsoft.azure/azure-mgmt-keyvault -->
+      <dependency>
+        <groupId>com.microsoft.azure</groupId>
+        <artifactId>azure-mgmt-keyvault</artifactId>
+        <version>1.11.1</version>
+      </dependency>
+      <dependency>
+        <groupId>com.microsoft.azure</groupId>
+        <artifactId>azure-client-runtime</artifactId>
+        <version>1.3.0</version>
+      </dependency>
+      <dependency>
+        <groupId>com.microsoft.azure</groupId>
+        <artifactId>azure-mgmt-resources</artifactId>
+        <version>1.3.0</version>
+      </dependency>
+      <dependency>
+        <groupId>com.microsoft.azure</groupId>
+        <artifactId>azure</artifactId>
+        <version>1.3.0</version>
+      </dependency>
 
+     <!-- Other Microsoft Dependencies -->
+      <dependency>
+        <groupId>com.microsoft.rest</groupId>
+        <artifactId>client-runtime</artifactId>
+        <version>1.6.1</version>
+      </dependency>
+
+    
+      <!-- Jackson Dependencies -->
       <dependency>
         <groupId>com.fasterxml.jackson.core</groupId>
         <artifactId>jackson-core</artifactId>
-        <version>2.7.2</version>
+        <version>${jackson.version}</version>
       </dependency>
       <dependency>
         <groupId>com.fasterxml.jackson.core</groupId>
         <artifactId>jackson-databind</artifactId>
-        <version>2.7.2</version>
+        <version>${jackson.version}</version>
       </dependency>
       <dependency>
         <groupId>com.fasterxml.jackson.core</groupId>
         <artifactId>jackson-annotations</artifactId>
-        <version>2.9.6</version>
+        <version>${jackson.version}</version>
       </dependency>
+      
+      <!-- Other Dependencies -->
+      <dependency>
+        <groupId>org.apache.commons</groupId>
+        <artifactId>commons-lang3</artifactId>
+  		<version>3.4</version>
+  	  </dependency>
       <dependency>
         <groupId>commons-codec</groupId>
         <artifactId>commons-codec</artifactId>
@@ -114,15 +178,11 @@
         <version>1.2.4</version>
       </dependency>
       <dependency>
-        <groupId>com.microsoft.azure</groupId>
-        <artifactId>azure-client-runtime</artifactId>
-		<version>1.3.0</version>
-      </dependency>
-      <dependency>
         <groupId>com.google.guava</groupId>
         <artifactId>guava</artifactId>
         <version>20.0</version>
       </dependency>
+      
       <!-- Test dependencies -->
       <dependency>
         <groupId>junit</groupId>
@@ -136,7 +196,11 @@
         <version>1.4.0</version>
         <scope>test</scope>
       </dependency>
-
+      <dependency>
+        <groupId>org.bouncycastle</groupId>
+        <artifactId>bcprov-jdk15on</artifactId>
+        <version>1.59</version>
+      </dependency>
     </dependencies>
   </dependencyManagement>
 
@@ -316,6 +380,19 @@
         		</lifecycleMappingMetadata>
         	</configuration>
         </plugin>
+        <plugin>
+          <groupId>org.apache.maven.plugins</groupId>
+          <artifactId>maven-failsafe-plugin</artifactId>
+          <version>2.22.0</version>
+          <executions>
+            <execution>
+              <goals>
+                <goal>integration-test</goal>
+                <goal>verify</goal>
+              </goals>
+            </execution>
+          </executions>
+        </plugin>
       </plugins>
     </pluginManagement>
   </build>
@@ -368,5 +445,6 @@
     <module>./azure-keyvault-webkey</module>
     <module>./azure-keyvault-cryptography</module>
     <module>./azure-keyvault-extensions</module>
+    <module>./azure-keyvault-complete</module>
   </modules>
 </project>
